# speed up the process in LuceneVerifyingIndexOutput