--- a/docs/changelog/96670.yaml
+++ b/docs/changelog/96670.yaml
@@ -1,0 +1,4 @@
+pr: 96670
+summary: speed up the process in LuceneVerifyingIndexOutput
+area: Store
+type: enhancement