As mentioned in [#96614](https://github.com/elastic/elasticsearch/issues/96614), The feature will be helpful to speedup the file chunk write process when there are a lot of small files to be recovery or restored. The Comparison Experimentation Based on NIO FS show that the speedup is 10 times.

       * Comparison Experimentation
       * Before optimization: chunk size 512KB, write times 10000, total cost time: 270827ms, avg cost time: 27.0827ms
       * After optimization: chunk size 512KB, write times 10000, total cost time: 11603ms, avg cost time: 1.1603ms