@rhluo 2023-06-08
if we define thes writeByte process as functionA(), now we consider the writeBytes process.
The original solution is as follows:

	for (int i = 0; i < length; i++)
		functionA()
I think you mean wo can just do as follows:
	
	out.writeBytes() //direct write parts before checksum position
	for(int i = checksumPosition; i < length; i++)
		functionA()