@DaveCTurner 2023-06-12
I don't have much to add to my preceding review - I sort of see why this is correct but that's not enough for me. The existing code is already too complicated (I wouldn't accept it if it were something we were adding today) and this change makes it worse.

I will open a PR to clean this area up, and address https://github.com/elastic/elasticsearch/issues/96614, in the next few weeks.