@rhluo 2023-06-12
@DaveCTurner Thank you very much for taking up a lot of your time to help me review the code. I try to provide a more understandable solution. the new solution contains three main steps: 
1) fill checksum using array copy
2) verify when writtenBytes reached file end
3) write through by call out.writeBytes

Finally, thanks again anyway, I learned a lot from communicating with you. I will also contribute my ideas to ES community at the next opportunity.