@DaveCTurner 2023-06-21
I opened https://github.com/elastic/elasticsearch/pull/96975 (and marked @rhluo as a co-author) so I'm closing this.